Tax rate fix for bug ID 3313927. Changed tax rate fields in three tables in the database.sql file. You'll have to manually make the same changes or use the sql file to create a new set of tables and copy your data over.
Tom

git-svn-id: svn+ssh://jekkos@svn.code.sf.net/p/opensourcepos/code/@44 c3eb156b-1dc0-44e1-88ae-e38439141b53

+++ database/database.sql
@@ -163,7 +163,7 @@ CREATE TABLE `ospos_items` (
 CREATE TABLE `ospos_items_taxes` (
   `item_id` int(10) NOT NULL,
   `name` varchar(255) NOT NULL,
-  `percent` double(15,2) NOT NULL,
+  `percent` double(15,3) NOT NULL,
   PRIMARY KEY (`item_id`,`name`,`percent`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
@@ -406,7 +406,7 @@ CREATE TABLE `ospos_sales_items_taxes` (
   `item_id` int(10) NOT NULL,
   `line` int(3) NOT NULL DEFAULT '0',
   `name` varchar(255) NOT NULL,
-  `percent` double(15,2) NOT NULL,
+  `percent` double(15,3) NOT NULL,
   PRIMARY KEY (`sale_id`,`item_id`,`line`,`name`,`percent`),
   KEY `item_id` (`item_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
@@ -493,7 +493,7 @@ CREATE TABLE `ospos_sales_suspended_items_taxes` (
   `item_id` int(10) NOT NULL,
   `line` int(3) NOT NULL DEFAULT '0',
   `name` varchar(255) NOT NULL,
-  `percent` double(15,2) NOT NULL,
+  `percent` double(15,3) NOT NULL,
   PRIMARY KEY (`sale_id`,`item_id`,`line`,`name`,`percent`),
   KEY `item_id` (`item_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
